+++ src/PhpSpreadsheet/Writer/Xlsx/Worksheet.php
@@ -1602,7 +1602,7 @@ class Worksheet extends WriterPart
 
                     break;
                 case 's':            // String
-                    $this->writeCellString($objWriter, $mappedType, $cellValueString, $flippedStringTable);
+                    $this->writeCellString($objWriter, $mappedType, ($cellValue instanceof RichText) ? $cellValue : $cellValueString, $flippedStringTable);
 
                     break;
                 case 'f':            // Formula